+++ mobile/src/components/atoms/icon/index.ts
@@ -1,0 +1,4 @@
+export { Icon } from './Icon';
+export type { IconProps } from './Icon.types';
+export { ICON_SIZES, ICON_FAVORITE_GOLD } from './constants';
+export type { IconSizeName } from './constants';